Add swiftlint

+++ TestView/DataCore/CoreDataManager.swift
@@ -16,59 +16,45 @@ class CoreDataManager {
 
     func getEntityForName(_ string: String) -> NSEntityDescription {
         return NSEntityDescription.entity(forEntityName: string, in: persistentContainer.viewContext)!
-        
     }
-    
-    func countObjectRequest(entityName: String, filterKey: String) -> Int?{
+    func countObjectRequest(entityName: String, filterKey: String) -> Int? {
         let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: entityName)
         fetchRequest.predicate = NSPredicate(format: "name CONTAINS[cd] %@", filterKey)
-        do{
+        do {
             let count = try persistentContainer.viewContext.count(for: fetchRequest)
             return count
-        }catch{
+        } catch {
             print("error reqeust count")
         }
         return nil
     }
-    
-    func deleteObject(entityName: String, filterKey: String){
+    func deleteObject(entityName: String, filterKey: String) {
         let fetchRequestDel = NSFetchRequest<NSFetchRequestResult>(entityName: entityName)
         fetchRequestDel.predicate = NSPredicate(format: "name == %@", filterKey)
-        do{
-            let arrUsrObj = try persistentContainer.viewContext.fetch(fetchRequestDel)
-            for usrObj in arrUsrObj as! [NSManagedObject] {
+        do {
+            guard let arrUsrObj = try persistentContainer.viewContext.fetch(fetchRequestDel) as? [NSManagedObject]
+                else { return }
+            for usrObj in arrUsrObj {
                  persistentContainer.viewContext.delete(usrObj)
             }
-        }catch{
+        } catch {
             print("error reqeust delete")
         }
         print(persistentContainer.viewContext.deletedObjects)
-        
         saveContext()
     }
-    
-    func getFetchAllPerson(entityName: String) -> [Any]?{
+    func getFetchAllPerson(entityName: String) -> [Any]? {
         let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: entityName)
-        do{
+        do {
             let objPerson = try persistentContainer.viewContext.fetch(fetchRequest)
             return objPerson
-        }catch{
+        } catch {
             print("error reqeust all person")
         }
         return nil
     }
 
     // MARK: - Core Data stack
-    func getFetchResultsController(entityName: String, sortDescriptorKey: String, filterKey: String?) -> NSFetchedResultsController<NSFetchRequestResult> {
-        let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: entityName)
-        let sortDescriptor = NSSortDescriptor(key: sortDescriptorKey, ascending: true)
-        fetchRequest.sortDescriptors = [sortDescriptor]
-        if let filter = filterKey {
-            fetchRequest.predicate = NSPredicate(format: "Person.name = %@", filter)
-        }
-        let fetchedResultsVc = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: persistentContainer.viewContext, sectionNameKeyPath: nil, cacheName: nil)
-        return fetchedResultsVc
-    }
 
     lazy var persistentContainer: NSPersistentContainer = {
         /*
@@ -78,19 +64,8 @@ class CoreDataManager {
          error conditions that could cause the creation of the store to fail.
         */
         let container = NSPersistentContainer(name: "ModelPerson")
-        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+        container.loadPersistentStores(completionHandler: { (_, error) in
             if let error = error as NSError? {
-                // Replace this implementation with code to handle the error appropriately.
-                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
-
-                /*
-                 Typical reasons for an error here include:
-                 * The parent directory does not exist, cannot be created, or disallows writing.
-                 * The persistent store is not accessible, due to permissions or data protection when the device is locked.
-                 * The device is out of space.
-                 * The store could not be migrated to the current model version.
-                 Check the error message to determine what the actual problem was.
-                 */
                 fatalError("Unresolved error \(error), \(error.userInfo)")
             }
         })
@@ -105,8 +80,6 @@ class CoreDataManager {
             do {
                 try context.save()
             } catch {
-                // Replace this implementation with code to handle the error appropriately.
-                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
                 let nserror = error as NSError
                 fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
             }
